S10P31A106-2 fix: SecureRandom 버그 수정

+++ backend/src/main/java/life/bareun/diary/habit/service/HabitServiceImpl.java
@@ -355,7 +355,7 @@ public class HabitServiceImpl implements HabitService {
         if (totalCount > 0) {
             SecureRandom secureRandom = new SecureRandom();
             while (totalCount > 0) {
-                Long randomNumber = secureRandom.nextLong(313L) + 1L;
+                Long randomNumber = secureRandom.nextInt(313) + 1L;
                 if (set.contains(randomNumber)) {
                     continue;
                 }
